# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: PV-DG Edge Local API
  version: 0.1.0
servers:
  - url: /api/v1
paths:
  /system/health:
    get:
      summary: Get local platform health
      tags: [System]
      responses:
        '200': { description: OK }
  /auth/setup-admin:
    post:
      summary: First-run admin setup
      tags: [Auth]
      responses:
        '201': { description: Admin created }
        '409': { description: Setup already complete }
  /auth/login:
    post:
      summary: Login with email and password
      tags: [Auth]
      responses:
        '200': { description: Access and refresh token issued }
        '401': { description: Invalid credentials }
  /auth/refresh:
    post:
      summary: Refresh access token
      tags: [Auth]
      responses:
        '200': { description: Access token issued }
        '401': { description: Invalid refresh token }
  /auth/logout:
    post:
      summary: Logout current session
      tags: [Auth]
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Logout acknowledged }
  /auth/me:
    get:
      summary: Get current user
      tags: [Auth]
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Current user }
  /roles:
    get:
      summary: List roles and permissions
      tags: [Roles]
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Roles }
  /telemetry/live/site/{siteId}:
    get:
      summary: Get latest live telemetry snapshot for a site from Redis cache
      tags: [Telemetry]
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/siteId'
      responses:
        '200':
          description: Latest site telemetry snapshot
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LiveTelemetryResponse'
        '401': { description: Authentication required }
        '403': { description: Permission denied }
        '404': { description: Live snapshot not found }
  /telemetry/live/device/{deviceId}:
    get:
      summary: Get latest live telemetry snapshot for a device from Redis cache
      tags: [Telemetry]
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/deviceId'
      responses:
        '200':
          description: Latest device telemetry snapshot
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LiveTelemetryResponse'
        '401': { description: Authentication required }
        '403': { description: Permission denied }
        '404': { description: Live snapshot not found }
  /sites:
    get:
      summary: List sites
      tags: [Sites]
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Sites }
    post:
      summary: Create site
      tags: [Sites]
      security: [{ bearerAuth: [] }]
      responses:
        '201': { description: Created }
  /sites/{siteId}:
    get:
      summary: Get site
      tags: [Sites]
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/siteId'
      responses:
        '200': { description: Site }
        '404': { description: Not found }
    patch:
      summary: Update site
      tags: [Sites]
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/siteId'
      responses:
        '200': { description: Updated }
        '404': { description: Not found }
    delete:
      summary: Delete site
      tags: [Sites]
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/siteId'
      responses:
        '200': { description: Deleted }
        '404': { description: Not found }
  /devices:
    get:
      summary: List devices
      tags: [Devices]
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Devices }
    post:
      summary: Create device
      tags: [Devices]
      security: [{ bearerAuth: [] }]
      responses:
        '201': { description: Created }
  /devices/{deviceId}:
    get:
      summary: Get device
      tags: [Devices]
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/deviceId'
      responses:
        '200': { description: Device }
        '404': { description: Not found }
    patch:
      summary: Update device
      tags: [Devices]
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/deviceId'
      responses:
        '200': { description: Updated }
        '404': { description: Not found }
    delete:
      summary: Delete device
      tags: [Devices]
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/deviceId'
      responses:
        '200': { description: Deleted }
        '404': { description: Not found }
  /devices/{deviceId}/test-connection:
    post:
      summary: Placeholder for future acquisition connection test
      tags: [Devices]
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/deviceId'
      responses:
        '501': { description: Scheduled for acquisition phase }
  /devices/{deviceId}/read-register:
    post:
      summary: Placeholder for future validated register read
      tags: [Devices]
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/deviceId'
      responses:
        '501': { description: Scheduled for acquisition phase }
  /drivers:
    get:
      summary: List drivers
      tags: [Drivers]
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Drivers }
  /drivers/import:
    post:
      summary: Import driver manifest metadata
      tags: [Drivers]
      security: [{ bearerAuth: [] }]
      responses:
        '201': { description: Imported }
  /drivers/{driverId}:
    get:
      summary: Get driver
      tags: [Drivers]
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/driverId'
      responses:
        '200': { description: Driver }
        '404': { description: Not found }
  /drivers/{driverId}/validate:
    post:
      summary: Validate stored driver metadata
      tags: [Drivers]
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/driverId'
      responses:
        '200': { description: Validation result }
  /drivers/{driverId}/export:
    get:
      summary: Export stored driver metadata
      tags: [Drivers]
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/driverId'
      responses:
        '200': { description: Driver export }
components:
  parameters:
    siteId:
      name: siteId
      in: path
      required: true
      schema: { type: string, format: uuid }
    deviceId:
      name: deviceId
      in: path
      required: true
      schema: { type: string, format: uuid }
    driverId:
      name: driverId
      in: path
      required: true
      schema: { type: string, format: uuid }
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    LiveTelemetryResponse:
      type: object
      required: [success, data, meta]
      properties:
        success: { type: boolean, example: true }
        data:
          type: object
          required: [updatedAt, payload]
          properties:
            siteId: { type: string, format: uuid }
            deviceId: { type: string, format: uuid }
            updatedAt: { type: string, format: date-time }
            payload:
              type: object
              description: Canonical telemetry payload validated from MQTT.
        meta:
          type: object
          properties:
            requestId: { type: string }